Split auto-graphql stuff out ouf pouch module
- Add new src/workspace/graphql/ module
- Move graphql stuff from pouch to there
- Specify definitions for pouch separately from definitions for graphql
- Specify that concrete definitions include both pouch and graphql stuff
